Implement sorting of articles.
There are four sorts: views and title in both increasing and decreasing as well as no sort.
These are implemented in comparisons.js and imported into a Map for access.
To get increasing and decreasing, the function is only defined once and can be multiplied by -1 to revert the order.